### 2023-06-02 #5516 @sampottinger
Hello! Excited about this and thanks for taking us to this step. Please note that pyopensci had one more thing for us to change in our badges and the current version is 1.0.2. Thanks!

See https://github.com/pyOpenSci/software-submission/issues/93
