### 2023-06-04 #5516 @sampottinger
Oh ok :sweat_smile:. Sorry I might need someone's help with updating that version. In the meantime, I'll take care of that DOI. I also know I'm just a touch over the recommended length so will try to shorten and I read some recent JOSS reviews here so will try to take care of a few small things I've seen highlighted elsewhere.
