### 2023-06-21 #5516 @graciellehigino
Hi all! This seems like a really interesting submission! However, I don't think I can help with this one because I always have a hard time finding Python users among ecologists. I think I'm on the wrong network for this... Sorry!
