### 2023-06-26 #5593 @editorialbot
[REVIEW]: Pyafscgap.org: Open source multi-modal Python-based tools for NOAA AFSC RACE GAP
**Submitting author:** @sampottinger (<a href="http://orcid.org/0000-0002-0458-4985">A Samuel Pottinger</a>)
**Repository:** https://github.com/SchmidtDSE/afscgap
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.4
