### 2023-06-26 #5593 @Kevin-Mattheus-Moerman
@sampottinger please try to address the potentially missing DOI listed :point_up: , you can use `@editorialbot check references` to check the DOIs again.
