### 2023-06-26 #5593 @Kevin-Mattheus-Moerman
@sampottinger I have checked your paper and noticed some minor typos, please address them, see below. I also recommend that you proof read your paper once more and amend where needed.

- [ ] Please add  `California, United States of America` to your affiliation
- [ ] Fix `compliation` -> `compilation`
- [ ] Fix `compatability` -> `compatibility`
- [ ] Fix `speices` -> `species`
- [ ] Fix `synchoronous` -> `synchronous`
