### 2023-06-26 #5593 @sampottinger
Thanks for all of this @Kevin-Mattheus-Moerman!

> some minor typos

Thank you for flagging those I’ll address shortly.

> Relating to the ZENODO archive:

I’ll confirm these items shortly!

> please try to address the potentially missing DOI listed

Thanks! ORDS does not have a DOI as far as I can tell. The one suggested is a [book for Oracle Cloud Services](https://link.springer.com/chapter/10.1007/978-1-4842-6011-1_12). Please advise how you want to continue but the current citation may be best.
