### 2023-06-28 #5593 @sampottinger
Hello @Kevin-Mattheus-Moerman! Thank you for your comments. I have:

- Addressed the typos you found. Appreciate it!
- Updated our affiliations as you requested.
- Corrected the zenodo metadata as you indicated.
- Performed another read-through with minor expansion while still trying to remain close to the JOSS word count limits.

My apologies but can you please update the version to `1.0.4` (@editorialbot set 1.0.4 as version). Thanks!
