### 2023-06-29 #5593 @Kevin-Mattheus-Moerman
@sampottinger it would be better to update the version and archive DOI here. We should be able to do that.

@openjournals/dev @arfon can you help process the following changes for this paper:

`@editorialbot set 1.0.4 as version`
`@editorialbot set 10.5281/zenodo.8091843 as archive`
