### 2023-06-29 #5593 @xuanxu
@Kevin-Mattheus-Moerman: Running those two commands to update the version and the archive values and then `@editorialbot reaccept` should work.
